Started work on evidence. Bug fix with categories on add detailed transaction screen

+++ app/src/main/java/com/davidgrath/expensetracker/entities/db/EvidenceDb.kt
@@ -9,7 +9,7 @@ data class EvidenceDb(
     val id: Long?,
     val transactionId: Long,
     val type: Type,
-    val sizeBytes: Int?,
+    val sizeBytes: Long?,
     val sha256: String?,
     val mimeType: String?,
     val uri: String,
